terramod-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-hcl2==4.3.2
jinja2==3.1.3
pyyaml==6.0.1
pytest==7.4.4
python-multipart==0.0.6
